+++ src/main/java/de/codescape/jira/plugins/scrumpoker/ScrumPokerConstants.java
@@ -10,6 +10,11 @@ public final class ScrumPokerConstants {
      */
     public static final String SCRUM_POKER_PLUGIN_KEY = "de.codescape.jira.plugins.scrum-poker";
 
+    /**
+     * Active Objects table prefix calculated from the MD5 hash of the plugin key name.
+     */
+    public static final String ACTIVE_OBJECTS_TABLE_PREFIX = "AO_1FA2A8_";
+
     // Prevent utility class from being instantiated
     private ScrumPokerConstants() {
         throw new AssertionError("ScrumPokerConstants must not be instantiated.");