+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ShowStatisticsAction.java
@@ -3,7 +3,7 @@ package de.codescape.jira.plugins.scrumpoker.action;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
-import de.codescape.jira.plugins.scrumpoker.model.ScrumPokerStatistics;
+import de.codescape.jira.plugins.scrumpoker.model.statistics.ScrumPokerStatistics;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSessionService;
 import org.springframework.beans.factory.annotation.Autowired;