+++ src/main/java/de/codescape/jira/plugins/scrumpoker/model/statistics/ScrumPokerStatistics.java
@@ -1,4 +1,6 @@
-package de.codescape.jira.plugins.scrumpoker.model;
+package de.codescape.jira.plugins.scrumpoker.model.statistics;
+
+import java.util.List;
 
 /**
  * The {@link ScrumPokerStatistics} object holds all statistics that are queried and get displayed on the Scrum Poker
@@ -9,6 +11,8 @@ public class ScrumPokerStatistics {
     private long sessionsTotal;
     private long votesTotal;
     private long sessionsParticipated;
+    private List<VoteDistribution> globalVotesDistribution;
+    private List<VoteDistribution> personalVotesDistribution;
 
     /**
      * Set the total number of Scrum Poker sessions started.
@@ -50,6 +54,34 @@ public class ScrumPokerStatistics {
      */
     public long getSessionsParticipated() {
         return sessionsParticipated;
+    }
+
+    /**
+     * Set the global distribution of votes.
+     */
+    public void setGlobalVotesDistribution(List<VoteDistribution> globalVotesDistribution) {
+        this.globalVotesDistribution = globalVotesDistribution;
+    }
+
+    /**
+     * Return the global distribution of votes.
+     */
+    public List<VoteDistribution> getGlobalVotesDistribution() {
+        return globalVotesDistribution;
+    }
+
+    /**
+     * Set the personal distribution of votes.
+     */
+    public void setPersonalVotesDistribution(List<VoteDistribution> personalVotesDistribution) {
+        this.personalVotesDistribution = personalVotesDistribution;
+    }
+
+    /**
+     * Return the personal distribution of votes.
+     */
+    public List<VoteDistribution> getPersonalVotesDistribution() {
+        return personalVotesDistribution;
     }
 
 }